folded config.py into dataset.py
