+++ esp_data/__init__.py
@@ -1,5 +1,6 @@
 from .dataset import (
     Dataset,
+    DatasetConfig,
     DatasetInfo,
     dataset_from_config,
     list_registered_datasets,
@@ -11,6 +12,7 @@ __all__ = [
     "dataset_from_config",
     "Dataset",
     "DatasetInfo",
+    "DatasetConfig",
     "list_registered_datasets",
     "print_registered_datasets",
     "register_dataset",
